fix: handled multiple roles issue

+++ src/main/java/com/example/blog/controllers/CommentController.java
@@ -1,10 +1,17 @@
 package com.example.blog.controllers;
 
+import com.example.blog.config.SecurityUtils;
 import com.example.blog.dto.CommentDto;
+import com.example.blog.entity.Comment;
+import com.example.blog.entity.Post;
+import com.example.blog.entity.User;
+import com.example.blog.exceptions.ResourceNotFoundException;
+import com.example.blog.repository.CommentRepository;
 import com.example.blog.services.CommentService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -15,6 +22,13 @@ public class CommentController {
     @Autowired
     private CommentService commentService;
 
+    @Autowired
+    private CommentRepository commentRepository;
+
+    @Autowired
+    private SecurityUtils securityUtils;
+
+    @PreAuthorize("hasRole('SUPER_ADMIN') or hasRole('ADMIN') or hasRole('NORMAL')")
     @PostMapping("/post/{postId}/user/{userId}/comment")
     public ResponseEntity<CommentDto> createCOmment(@RequestBody CommentDto commentDto, @PathVariable Integer postId, @PathVariable Integer userId ){
         CommentDto response=this.commentService.createComment(commentDto,postId,userId);
@@ -23,6 +37,22 @@ public class CommentController {
 
     @DeleteMapping("/comments/{commentId}")
     public ResponseEntity<String> deleteComment(@PathVariable Integer commentId ){
+
+        User loggedInUser = securityUtils.getLoggedInUserDetails();
+        boolean isNormalUser = loggedInUser.getRoles().stream()
+                .anyMatch(role -> role.getName().equals("ROLE_NORMAL"));
+
+        Comment commentTOBeDeleted=this.commentRepository.findById(commentId).orElseThrow(()-> new ResourceNotFoundException("Comment","comment ID ",commentId));
+        Integer commentOwnerId = commentTOBeDeleted.getUser().getId();
+        //normal user-> can delete its own comment only not of others
+        if ((isNormalUser && (loggedInUser.getId()!=commentOwnerId))
+                && !securityUtils.hasRole("ROLE_ADMIN")
+                && !securityUtils.hasRole("ROLE_SUPER_ADMIN")
+        ) {
+            return new ResponseEntity<>("You can only delete your own comments.", HttpStatus.FORBIDDEN);
+        }
+
+        //admin, super-admin
         String response=this.commentService.deleteComment(commentId);
         return new ResponseEntity<>(response,HttpStatus.OK);
     }
@@ -37,6 +67,21 @@ public class CommentController {
         if(commentDto.getContent()==null){
             return new ResponseEntity<>("content cannot be null",HttpStatus.BAD_REQUEST);
         }
+
+        User loggedInUser = securityUtils.getLoggedInUserDetails();
+        boolean isNormalUser = loggedInUser.getRoles().stream()
+                .anyMatch(role -> role.getName().equals("ROLE_NORMAL"));
+
+        Comment commentTOBeUpdated=this.commentRepository.findById(commentId).orElseThrow(()-> new ResourceNotFoundException("Comment","comment ID ",commentId));
+        Integer commentOwnerId = commentTOBeUpdated.getUser().getId();
+        //normal user-> can delete its own comment only not of others
+        if ((isNormalUser && (loggedInUser.getId()!=commentOwnerId))
+                && !securityUtils.hasRole("ROLE_ADMIN")
+                && !securityUtils.hasRole("ROLE_SUPER_ADMIN")
+        ) {
+            return new ResponseEntity<>("You can only update your own comments.", HttpStatus.FORBIDDEN);
+        }
+
 
         String response=this.commentService.updateComment(commentDto,commentId,postId,userId);
         return new ResponseEntity<>(response,HttpStatus.OK);
